ui: storage: Fix concurrency issue in questions cache
Since getAllExplainers is async, it can be called concurrently (when we
block on the await), which results in duplicate API requests to the
all questions endpoint.

Fix this by slapping a promise in _cache.explainers so that subsequent
callers just return that promise (which will be resolved by the first
callee) rather than duplicating the work.